+++ veecle-os-test/src/execute.rs
@@ -19,7 +19,7 @@
 /// async fn incrementor(mut writer: Writer<'_, Data>, mut trigger: Reader<'_, Trigger>) -> Never {
 ///     loop {
 ///         trigger.wait_for_update().await;
-///         writer.modify(|data| {
+///         writer.modify(|mut data| {
 ///             *data = Some(data.map_or(Data(0), |data| Data(data.0 + 1)));
 ///         }).await;
 ///     }